Added a jdk9 profile and a module-info.java file, to be skipped in all profiles except jdk9. Due to https://issues.apache.org/jira/browse/MCOMPILER-294, test can not yet be compiled with Maven. Javadoc generation also fails, so the following command line have to be used:
mvn clean install -Pjdk9 -Dmaven.test.skip=true -Dmaven.javadoc.skip=true

In addition, as of June 26th 2017 the following changes need to be done in the pom.xml file in order to compile with JDK9. Those changes are not included in this commit since we assume that the need for them is temporary: <packaging>bundle</packaging> element need to be changed into <packaging>jar</packaging> since maven-bundle-plugin version 3.3.0 does not seem to be compatible with JDK9 (but maven-jar-plugin is okay), and the Javadoc attach-sources execution needs to be removed.