163: Suppressed unused formal parameter PMD warning.
The constructor here defines a parameter which is not used/assigned at all in the constructor's body, so PMD's suggestion is to remove it. However, in case we do so, the tests will fail. That is, this parameter is indeed not used in the constructor's body, but it's -properly- used in the tests, while calling the constructor of the class.

Hence, it is comprehensive to suppress it at method level here.